orchestrator/requirements.txt
faster-whisper
transformers
sentencepiece
soundfile
numpy
python-dotenv
tqdm
requests
